feat(core): add custom exceptions (NotFoundException, BadRequestException)

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,9 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Domain;
 using MediatR;
+using System.Linq.Expressions;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
 
@@ -17,15 +20,19 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
 
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
-        //Retrieve domain entity object
+        // Retrieve domain entity object
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
-        //Verify that record exists
+        // Verify that record exists
+        if (leaveTypeToDelete == null)
+        {
+            throw new NotFoundException(nameof(LeaveType), request.Id);
+        }
 
-        //Remove from database
+        // Remove from database
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
 
-        //Return record id
+        // Return record id
         return Unit.Value;
     }
 }
